refactor: route app storage error helpers
