+++ rustfs/src/app/object_usecase.rs
@@ -14,9 +14,7 @@
 
 //! Object application use-case contracts.
 
-use super::object_api_utils::to_s3s_etag;
 use super::quota::checker::QuotaChecker;
-use super::storageclass;
 // Performance metrics recording (with zero-copy-metrics integration)
 use super::ECStore;
 use super::s3_api::multipart::parse_list_parts_params;
@@ -28,9 +26,14 @@ use super::storage_api::concurrency::{
 };
 use super::storage_api::deadlock_detector;
 use super::storage_api::ecfs::FS;
+use super::storage_api::error::{
+    DiskError, Error as EcstoreError, StorageError, is_all_buckets_not_found, is_err_bucket_not_found, is_err_object_not_found,
+    is_err_version_not_found,
+};
 use super::storage_api::head_prefix::{head_prefix_not_found_message, probe_prefix_has_children};
 use super::storage_api::helper::{OperationHelper, spawn_background_with_context};
 use super::storage_api::io::{DynReader, HashReader, WritePlan, compression_metadata_value, wrap_reader};
+use super::storage_api::object_utils::to_s3s_etag;
 use super::storage_api::options::{
     copy_dst_opts, copy_src_opts, del_opts, extract_metadata, extract_metadata_from_mime_with_object_name,
     filter_object_metadata, get_content_sha256_with_query, get_opts, normalize_content_encoding_for_storage, put_opts,
@@ -42,6 +45,7 @@ use super::storage_api::sse::{
     encryption_material_to_metadata, extract_server_side_encryption_from_headers, extract_ssec_params_from_headers,
     extract_ssekms_context_from_headers, get_buffer_size_opt_in, map_get_object_reader_error, sse_decryption, sse_encryption,
 };
+use super::storage_api::storage_class as storageclass;
 use super::storage_api::timeout_wrapper::{GetObjectTimeoutPolicy, RequestTimeoutWrapper};
 use super::storage_api::{
     RFC1123, check_preconditions, get_validated_store, has_replication_rules, parse_object_lock_legal_hold,
@@ -50,8 +54,6 @@ use super::storage_api::{
     validate_sse_headers_for_write, validate_ssec_for_read, wrap_response_with_cors,
 };
 use super::{AppReplicationConfigExt as _, AppVersioningConfigExt as _, predict_lifecycle_expiration, validate_restore_request};
-use super::{DiskError, is_all_buckets_not_found};
-use super::{Error as EcstoreError, StorageError, is_err_bucket_not_found, is_err_object_not_found, is_err_version_not_found};
 use super::{
     lifecycle::{
         bucket_lifecycle_audit::LcEventSrc,
